split code for equipments part
