+++ internal/utils/files.go
@@ -1,0 +1,74 @@
+package utils
+
+import (
+	"bytes"
+	"fmt"
+	"io"
+	"net/url"
+	"os"
+	"time"
+
+	"github.com/pkg/sftp"
+	"golang.org/x/crypto/ssh"
+)
+
+func GetFile(uri url.URL, connectionTimeout time.Duration) (io.Reader, error) {
+	if uri.Scheme == "sftp" {
+		return GetFileWithSftp(uri, connectionTimeout)
+	} else if uri.Scheme == "file" {
+		return GetFileWithFS(uri)
+	} else {
+		return nil, fmt.Errorf("Unsupported protocols %s", uri.Scheme)
+	}
+
+}
+
+func GetFileWithFS(uri url.URL) (io.Reader, error) {
+	file, err := os.Open(uri.Path)
+	if err != nil {
+		return nil, err
+	}
+	defer file.Close()
+	var buffer bytes.Buffer
+	if _, err = buffer.ReadFrom(file); err != nil {
+		return nil, err
+	}
+	return &buffer, nil
+}
+
+func GetFileWithSftp(uri url.URL, connectionTimeout time.Duration) (io.Reader, error) {
+	password, _ := uri.User.Password()
+	sshConfig := &ssh.ClientConfig{
+		User: uri.User.Username(),
+		Auth: []ssh.AuthMethod{
+			ssh.Password(password),
+		},
+		HostKeyCallback: ssh.InsecureIgnoreHostKey(), //nolint:gosec
+		Timeout:         connectionTimeout,
+	}
+
+	sshClient, err := ssh.Dial("tcp", uri.Host, sshConfig)
+	if err != nil {
+		return nil, err
+	}
+	defer sshClient.Close()
+
+	// open an SFTP session over an existing ssh connection.
+	client, err := sftp.NewClient(sshClient)
+	if err != nil {
+		return nil, err
+	}
+	defer client.Close()
+
+	file, err := client.Open(uri.Path)
+	if err != nil {
+		return nil, err
+	}
+	defer file.Close()
+	var buffer bytes.Buffer
+	if _, err = file.WriteTo(&buffer); err != nil {
+		return nil, err
+	}
+	return &buffer, nil
+
+}
